pyproject.toml
[project]
name = "obsidian-mcp"
version = "0.1.0"
description = "MCP server for Obsidian vault operations via Local REST API"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp>=1.0.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "anyio[trio]",
    "pytest-anyio",
    "ruff>=0.3.0",
    "ipykernel",
]

[project.scripts]
obsidian-mcp = "obsidian_mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py312"

# Linting rules
lint.select = [
    "E",     # pycodestyle errors
    "W",     # pycodestyle warnings
    "F",     # pyflakes
    "I",     # isort
    "N",     # pep8-naming
    "UP",    # pyupgrade
    "B",     # flake8-bugbear
    "C4",    # flake8-comprehensions
    "SIM",   # flake8-simplify
]
lint.ignore = [
    "E501",  # line too long (handled by formatter)
]

# Formatting
format.quote-style = "double"
format.indent-style = "space"
